% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_txt.R
\name{clean.text}
\alias{clean.text}
\title{Clean text files}
\usage{
clean.text(in_dir)
}
\arguments{
\item{in_dir}{directory with input text files}
}
\value{
list of cleaned texts
}
\description{
Clean text files
}
\examples{
clean.text(in_dir = "./inst/extdata/test_pdfs/")
}
