% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kw.R
\docType{data}
\name{kw}
\alias{kw}
\title{List of all keywords}
\format{
A vector of 121 character strings
}
\source{
Helmstetter, 2020
}
\usage{
kw
}
\description{
A vector of all keywords of interest
}
\keyword{datasets}
