% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lemma_data.R
\docType{data}
\name{lemma_data}
\alias{lemma_data}
\title{Lemma data}
\format{
A data frame with 47366 rows and 2 variables:
\describe{
  \item{inflected_form}{string of inflected forms}
  \item{lemma}{list of lemmas (roots)}
  ...
}
}
\source{
Helmstetter, 2020
}
\usage{
lemma_data
}
\description{
Lemma data (dictionary)
}
\keyword{datasets}
