% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi_auto.R
\name{semi.auto.paired}
\alias{semi.auto.paired}
\title{Interactively retrieve paired qualitative data from a paper}
\usage{
semi.auto.paired(
  in_dir,
  keywords1,
  keywords2,
  highlight,
  n,
  cleaned_text,
  sorted_words
)
}
\arguments{
\item{in_dir}{directory with input text files}

\item{keywords1}{A vector of keywords as characters (i.e. traits of interest) in a vector}

\item{keywords2}{A second vector of keywords that must also be in the extracted text (exact matches only)}

\item{highlight}{A vector of keywords that are highlighted in the printed text}

\item{n}{A number specifying the threshold to consider keyword present}

\item{cleaned_text}{output of 'clean.text' - a list of cleaned text files}

\item{sorted_words}{output of 'generate.term.dataset' - a list of frequency of searched words mentioned in each paper}
}
\value{
A data frame containing matches above threshold and verified matches
}
\description{
Interactively retrieve paired qualitative data from a paper
}
\examples{
download.file("https://github.com/ajhelmstetter/papieRmache/raw/master/inst/extdata/test_pdfs.zip", destfile = "./test_pdfs.zip")
unzip("./test_pdfs.zip")

ct<-clean.text(in_dir = "./test_pdfs",all_keywords=kw)

keywords1<-c('diversification')
keywords2<-c('increase','decrease')

td<-generate.term.dataset(cleaned_text = ct, in_dir = "./testpdfs/",keywords = keywords1)

semi.auto.paired(in_dir = "./test_pdfs/", keywords1 = keywords1, keywords2 = keywords2, highlight = c('increase','decrease'), n = 100, cleaned_text = ct, sorted_words = td)

}
