% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi_auto.R
\name{semi.auto.value}
\alias{semi.auto.value}
\title{Interactively retrieve quantitative data from a paper}
\usage{
semi.auto.value(in_dir, keywords, out_file, cleaned_text, sorted_words)
}
\arguments{
\item{in_dir}{directory with input text files}

\item{keywords}{A set of keywords as characters (i.e. traits of interest) in a vector}

\item{out_file}{A name for the output file (RDS format)}

\item{cleaned_text}{output of 'clean.text' - a list of cleaned text files}

\item{sorted_words}{output of 'generate.term.dataset' - a list of frequency of searched words mentioned in each paper}
}
\value{
A data frame containing matches above threshold and verified matches
}
\description{
Interactively retrieve quantitative data from a paper
}
\examples{
keywords<-c('species')
ct<-clean.text(in_dir = "./inst/extdata/test_pdfs/")
td<-generate.term.dataset(cleaned_text = ct, in_dir = "./inst/extdata/testpdfs/",keywords = keywords)
semi.auto.value(in_dir = "./inst/extdata/test_pdfs/", keywords ="species", out_file ="species_number", cleaned_text = ct, sorted_words = td)

}
