% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopwords_extended.R
\docType{data}
\name{stopwords_extended}
\alias{stopwords_extended}
\title{List of stopwords}
\format{
A vector of 565 character strings
}
\source{
Helmstetter, 2020
}
\usage{
stopwords_extended
}
\description{
A vector of stopwords
}
\keyword{datasets}
