% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_txt.R
\name{clean.text}
\alias{clean.text}
\title{Clean text files}
\usage{
clean.text(in_dir, all_keywords)
}
\arguments{
\item{in_dir}{directory with input text files}

\item{all_keywords}{vector with all keywords you are interested in}
}
\value{
list of cleaned texts
}
\description{
Clean text files
}
\examples{

download.file("https://github.com/ajhelmstetter/papieRmache/raw/master/inst/extdata/test_pdfs.zip", destfile = "./test_pdfs.zip")
unzip("./test_pdfs.zip")
clean.text(in_dir = "./test_pdfs/",all_keywords=kw)
}
