% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_txt.R
\name{clean.text}
\alias{clean.text}
\title{Clean text files}
\usage{
clean.text(in_dir, all_keywords)
}
\arguments{
\item{in_dir}{directory with input text files}

\item{all_keywords}{vector with all keywords you are interested in}
}
\value{
list of cleaned texts
}
\description{
Clean text files
}
\examples{
data(papieRmache)
clean.text(in_dir = "./extdata/test_pdfs/",all_keywords=kw)
}
