% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_term_dataset.R
\name{generate.term.dataset}
\alias{generate.term.dataset}
\title{Generate frequencies of terms for each paper}
\usage{
generate.term.dataset(cleaned_text, in_dir, keywords)
}
\arguments{
\item{cleaned_text}{output of 'clean.text' - a list of cleaned text files}

\item{in_dir}{directory with input text files}

\item{keywords}{A set of keywords as characters (i.e. traits of interest) in a vector}
}
\value{
A list of frequencies of keywords for each paper
}
\description{
Generate frequencies of terms for each paper
}
\examples{
ct<-clean.text(in_dir = "./inst/extdata/test_pdfs/")
generate.term.dataset(cleaned_text = ct, in_dir = "./inst/extdata/testpdfs/",keywords = c("bisse","musse"))

}
