% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{semi.auto}
\alias{semi.auto}
\title{Interactively retrieve qualitative data from a paper}
\usage{
semi.auto(in_dir, keywords, n, out_file, cleaned_text, sorted_words)
}
\arguments{
\item{in_dir}{directory with input text files}

\item{keywords}{A set of keywords as characters (i.e. traits of interest) in a vector}

\item{n}{A number specifying the threshold to consider keyword present}

\item{out_file}{A name for the output file (RDS format)}

\item{cleaned_text}{output of 'clean.text' - a list of cleaned text files}

\item{sorted_words}{output of 'generate.term.dataset' - a list of frequency of searched words mentioned in each paper}
}
\value{
A data frame containing matches above threshold and verified matches
}
\description{
Interactively retrieve qualitative data from a paper
}
\examples{
keywords<-c('tree','shrub','epiphyte')
semi.auto(in_dir = "./pdfs/", keywords = keywords, n = 5, out_file = "growth_form", cleaned_text = cleaned_text, sorted_words = sorted_words)

}
