% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{semi.auto.value}
\alias{semi.auto.value}
\title{Interactively retrieve quantitative data from a paper}
\usage{
semi.auto.value(in_dir, keywords, out_file, cleaned_text, sorted_words)
}
\arguments{
\item{in_dir}{directory with input text files}

\item{keywords}{A set of keywords as characters (i.e. traits of interest) in a vector}

\item{out_file}{A name for the output file (RDS format)}

\item{cleaned_text}{output of 'clean.text' - a list of cleaned text files}

\item{sorted_words}{output of 'generate.term.dataset' - a list of frequency of searched words mentioned in each paper}
}
\value{
A data frame containing matches above threshold and verified matches
}
\description{
Interactively retrieve quantitative data from a paper
}
\examples{
semi.auto.value(in_dir = "./pdfs/", keywords ="species", out_file ="species_number", cleaned_text = cleaned_text, sorted_words = sorted_words)

}
